+++ src/elements/Character State Machines/Enemies/Slime.ts
@@ -1,10 +1,11 @@
 import { BodyHandle } from "engine/BodyHandle";
 import { Helpers } from "engine/Helpers";
 
-import { CharacterStates, EnemyAlertState, EnemyChaseState, StateTransitionManager, StateTransitionRule, characterPatrolState } from "elements/Character State Machines/CharacterStates";
+import { CharacterStates, StateTransitionManager, StateTransitionRule, characterPatrolState } from "elements/Character State Machines/CharacterStates";
 import { UpdateHandler } from "engine/MessageHandlers";
 import { Enemy } from "./Enemy";
 import { MODscriptThrowState } from "../MODscriptStates";
+import { EnemyChaseState, EnemyAlertState } from "../EnemyStates";
 
 
 class SlimePatrol extends characterPatrolState {
